Changed Date to Timestamp / fixed DateConverter
Changed java.sql.Date to java.sql.Timestamp, which additionally handles
the time. The getter and setter for the createdOn field returns or saves
a java.util.Date object, respectively.

The DateConverter now returns the correct time in the 'today' and
'yesterday' case. These two words are read from the message bundle.